[FE][shelter URL paths] use shelterId from URL for add pet submission
