+++ contract/shelter/addNewPet.ts
@@ -4,5 +4,5 @@ import { Methods } from "../methods";
 export namespace addNewPetApiDomain {
 	export const endpoint = "api/animal";
 	export const method = Methods.POST;
-	export interface requestBody extends NewAnimalAttributes {}
+	export type requestBody = NewAnimalAttributes;
 }
