+++ frontend/components/shelter/pet/add/Form.tsx
@@ -5,28 +5,18 @@ import Title from "antd/lib/typography/Title";
 import { PetApi } from "api/petApi";
 import { NewAnimal } from "common/types";
 import dayjs from "dayjs";
-import { Formik, FormikHelpers } from "formik";
+import { Formik } from "formik";
 import moment from "moment";
+import { useRouter } from "next/router";
 import React from "react";
 import styled from "styled-components";
-import {
-	array,
-	boolean,
-	date,
-	mixed,
-	number,
-	object,
-	SchemaOf,
-	string,
-	StringSchema,
-} from "yup";
 import { FormSection } from "./FormComponents";
+import { schema } from "./schema";
 
 // =============================================================================
 // Schema
 // =============================================================================
 const initialPet: NewAnimal = {
-	shelterId: "test", //TODO: revert to "" after testing
 	adoptionStatus: Animal.AdoptionStatus.Healthy,
 	species: Animal.Species.Cat,
 	name: "",
@@ -48,38 +38,6 @@ const initialPet: NewAnimal = {
 	visible: true,
 	animalImages: [],
 };
-const imageSchema: SchemaOf<Animal.Image> = object().shape({
-	thumbnailUrl: string().url().required(),
-	photoUrl: string().url().required(),
-});
-
-const schema: SchemaOf<NewAnimal> = object().shape({
-	shelterId: string().required(),
-	adoptionStatus: mixed<Animal.AdoptionStatus>()
-		.oneOf(Object.values(Animal.AdoptionStatus))
-		.required(),
-	species: mixed<Animal.Species>()
-		.oneOf(Object.values(Animal.Species))
-		.required(),
-	name: string().required(),
-	description: string() as StringSchema<string>, //TODO: revert to string().required() after testing and if added to form
-	healthIssues: string().required(),
-	gender: mixed<"M" | "F">().oneOf(["M", "F"]).required(),
-	dateOfBirth: date().required(),
-	sizeCm: number().nullable().defined(), //TODO: set validation if added to form
-	breed: string().required(),
-	color: string().required(),
-	weightKg: number().nullable().defined(), //TODO: set validation if added to form
-	furLength: string().required(),
-	vaccinated: boolean().nullable().defined(), //TODO: set validation if added to form
-	dewormed: boolean().nullable().defined(), //TODO: set validation if added to form
-	sterilised: boolean().nullable().defined(),
-	toiletTrained: boolean().required(),
-	adoptionFee: number().nullable().defined(), //TODO: set validation if added to form
-	intakeDate: date().required(),
-	visible: boolean().required(),
-	animalImages: array().of(imageSchema.required()).required().min(1),
-});
 
 // =============================================================================
 // Helpers
@@ -94,18 +52,18 @@ const dateToDateString = (date: Date | null) => {
 // Form
 // =============================================================================
 export const AddPetForm = () => {
-	const handleSubmit = async (
-		values: NewAnimal,
-		actions: FormikHelpers<NewAnimal>,
-	) => {
+	const router = useRouter();
+	const shelterId = router.query.shelterId as string;
+
+	const handleSubmit = async (values: NewAnimal) => {
 		const transformedValues: Shelter.addNewPetApiDomain.requestBody = {
 			...values,
 			dateOfBirth: dateToDateString(values.dateOfBirth),
 			intakeDate: dateToDateString(values.intakeDate)!,
+			shelterId,
 		};
 		/*
 			TODO:
-			Append shelter ID from url/login context
 			missing form inputs -
 				adoption fee
 				description
